feat(lane_change): modify lane change target boundary check to consider velocity (#8961)
* check if candidate path footprint exceeds target lane boundary when lc velocity is above minimum


* move functions to relevant module


* suppress unused function cppcheck


* minor change


---------
