+++ planning/behavior_path_planner/autoware_behavior_path_avoidance_by_lane_change_module/src/interface.cpp
@@ -44,11 +44,14 @@ AvoidanceByLaneChangeInterface::AvoidanceByLaneChangeInterface(
 {
 }
 
+// cppcheck-suppress unusedFunction
 bool AvoidanceByLaneChangeInterface::isExecutionRequested() const
 {
   return module_type_->isLaneChangeRequired() && module_type_->specialRequiredCheck() &&
          module_type_->isValidPath();
 }
+
+// cppcheck-suppress unusedFunction
 void AvoidanceByLaneChangeInterface::processOnEntry()
 {
   waitApproval();
